dev-requirements.txt
codacy-coverage
pytest==3.2.3
pytest-cov==2.5.1
pytest-flask==0.10.0